Add CLA template documents.
Thanks to harmonyagreements.org.  We require contributions to this
repository to be accompanied by a CLA, so that we may include them
in the commercially licensed versions of this code.  Note that the CLA
assures that we will continue to distribute contributions under the
same license as this repository in addition to any other distribution.